Runtime Improvements
Removed most of the randomization from the algorithm. The algorithm
will run much faster, but it spends less time trying to reduce the
number of keys to be harvested, and the output is no longer guaranteed
to get the truly optimal number of fields.

Input files are now semicolon delimited and can state lat, long
directly or can have the Intel URL in their place. See README.md.

Patched a massive memory leak in branch-and-bound algorithm.

Other minor improvements